requirements.txt
discord.py==1.6.0
pytz==2020.1
discord==1.0.1
python-decouple==3.4
